Split up api tests into files
