+++ include/test/crossPlatformTestRunner.hpp
@@ -5,7 +5,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-#ifndef PROS_VERSION_MAJOR
 #ifndef _OKAPI_CROSSPLATFORMTESTRUNNER_HPP_
 #define _OKAPI_CROSSPLATFORMTESTRUNNER_HPP_
 
@@ -49,5 +48,4 @@ namespace okapi {
     void test_print_report();
 } // namespace okapi
 
-#endif
 #endif
