+++ include/test/tests/api/allApiTests.hpp
@@ -1,0 +1,19 @@
+/**
+ * @author Ryan Benasutti, WPI
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+#ifndef _OKAPI_ALLAPITESTS_HPP_
+#define _OKAPI_ALLAPITESTS_HPP_
+
+#include "test/tests/api/utilTests.hpp"
+#include "test/tests/api/filterTests.hpp"
+#include "test/tests/api/buttonTests.hpp"
+#include "test/tests/api/controlTests.hpp"
+#include "test/tests/api/chassisModelTests.hpp"
+#include "test/tests/api/chassisControllerTests.hpp"
+#include "test/crossPlatformTestRunner.hpp"
+
+#endif
