+++ include/test/tests/api/buttonTests.hpp
@@ -1,0 +1,13 @@
+/**
+ * @author Ryan Benasutti, WPI
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+#ifndef _OKAPI_BUTTONTESTS_HPP_
+#define _OKAPI_BUTTONTESTS_HPP_
+
+void testButtons();
+
+#endif
