+++ test/crossPlatformTestRunner.cpp
@@ -5,7 +5,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-#ifndef PROS_VERSION_MAJOR
 #include "test/crossPlatformTestRunner.hpp"
 
 namespace okapi {
@@ -42,5 +41,3 @@ void test_print_report() {
   }
 }
 } // namespace okapi
-
-#endif
