+++ test/testMain.cpp
@@ -1,0 +1,23 @@
+/**
+ * @author Ryan Benasutti, WPI
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+#include "test/tests/api/allApiTests.hpp"
+
+int main() {
+  using namespace okapi;
+
+  testChassisScales();
+  testChassisModels();
+  testControlUtils();
+  testButtons();
+  testFilters();
+  testUtil();
+
+  test_print_report();
+
+  return 0;
+}
